requirements.txt
BeautifulSoup==3.2.1
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.11.5
chardet==3.0.4
fake-useragent==0.1.8
html2text==2017.10.4
html5lib==1.0.1
idna==2.6
lxml==4.1.1
mechanize==0.3.6
pkg-resources==0.0.0
python-slugify==1.2.4
request==0.0.22
requests==2.18.4
six==1.11.0
Unidecode==0.4.21
urllib3==1.22
webencodings==0.5.1
